requirements.txt
cryptography~=41.0.3
pip~=22.3.1
pytz~=2023.3
Django~=4.2.3
asgiref~=3.7.2
sqlparse~=0.4.4
requests~=2.31.0
django-allauth~=0.55.0
PyJWT~=2.8.0
certifi~=2023.7.22
oauthlib~=3.2.2
idna~=3.4
urllib3~=2.0.4
pycparser~=2.21
setuptools~=65.5.0
django-filter~=23.2
celery~=4.4.1